# backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Aspire NGO Platform API",
    "description": "Team Aspire — donations, projects, LMS, CRM, CMS and operations API",
    "version": "1.0.0"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "securityDefinitions": {
    "token": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "paths": {
...
    "/api/v1/cms/pages": {
      "get": {
        "tags": [
          "CMS"
        ],
        "summary": "List pages",
        "description": "All CMS pages",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "data": {
                  "type": "object"
                },
                "message": {
                  "type": "string",
                  "example": "Success"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "error"
                },
                "data": {
                  "type": "object"
                },
                "message": {
                  "type": "string",
                  "example": "Bad Request"
                },
                "errors": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "error"
                },
                "data": {
                  "type": "object"
                },
                "message": {
                  "type": "string",
                  "example": "Not Found"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "error"
                },
                "data": {
                  "type": "object"
                },
                "message": {
                  "type": "string",
                  "example": "Internal Server Error"
                }
              }
            }
          }
        },
        "parameters": [],
        "operationId": "ListPagesController"
      },
      "post": {
        "tags": [
          "CMS"
        ],
        "summary": "Create page",
        "description": "New block-based page",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "data": {
                  "type": "object"
                },
                "message": {
                  "type": "string",
                  "example": "Success"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "error"
                },
                "data": {
                  "type": "object"
                },
                "message": {
                  "type": "string",
                  "example": "Bad Request"
                },
                "errors": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "error"
                },
                "data": {
                  "type": "object"
                },
                "message": {
                  "type": "string",
                  "example": "Not Found"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "error"
                },
                "data": {
                  "type": "object"
                },
                "message": {
                  "type": "string",
                  "example": "Internal Server Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string",
                  "minLength": 2,
                  "maxLength": 160
                },
                "slug": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 160
                },
                "blocks": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "enum": [
                          "hero",
                          "richText",
                          "stats",
                          "cta",
                          "faq"
                        ]
                      },
                      "heading": {
                        "type": "string"
                      },
                      "subheading": {
                        "type": "string"
                      },
                      "body": {
                        "type": "string"
                      },
                      "ctaLabel": {
                        "type": "string"
                      },
                      "ctaHref": {
                        "type": "string"
                      },
                      "image": {
                        "type": "string"
                      },
                      "items": {
                        "type": "array",
                        "items": {}
                      }
                    },
                    "required": [
                      "type"
                    ],
                    "additionalProperties": false
                  }
                },
                "seo": {
                  "type": "object",
                  "properties": {
                    "title": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    }
                  },
                  "additionalProperties": false
                }
              },
              "required": [
                "title"
              ],
              "additionalProperties": false
            }
          }
        ],
        "operationId": "CreatePageController"
      }
    },
    "/api/v1/cms/pages/{id}": {
      "get": {
        "tags": [
          "CMS"
        ],
        "summary": "Get page",
        "description": "Page with blocks",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "data": {
                  "type": "object"
                },
                "message": {
                  "type": "string",
                  "example": "Success"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "error"
                },
                "data": {
                  "type": "object"
                },
                "message": {
                  "type": "string",
                  "example": "Bad Request"
                },
                "errors": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "error"
                },
                "data": {
                  "type": "object"
                },
                "message": {
                  "type": "string",
                  "example": "Not Found"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "error"
                },
                "data": {
                  "type": "object"
                },
                "message": {
                  "type": "string",
                  "example": "Internal Server Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "operationId": "GetPageController"
      },
      "patch": {
        "tags": [
          "CMS"
        ],
        "summary": "Update page",
        "description": "Edit title/slug/blocks/SEO",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "data": {
                  "type": "object"
                },
                "message": {
                  "type": "string",
                  "example": "Success"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "error"
                },
                "data": {
                  "type": "object"
                },
                "message": {
                  "type": "string",
                  "example": "Bad Request"
                },
                "errors": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "error"
                },
                "data": {
                  "type": "object"
                },
                "message": {
                  "type": "string",
                  "example": "Not Found"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "error"
                },
                "data": {
                  "type": "object"
                },
                "message": {
                  "type": "string",
                  "example": "Internal Server Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string",
                  "minLength": 2,
                  "maxLength": 160
                },
                "slug": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 160
                },
                "blocks": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "enum": [
                          "hero",
                          "richText",
                          "stats",
                          "cta",
                          "faq"
                        ]
                      },
                      "heading": {
                        "type": "string"
                      },
                      "subheading": {
                        "type": "string"
                      },
                      "body": {
                        "type": "string"
                      },
                      "ctaLabel": {
                        "type": "string"
                      },
                      "ctaHref": {
                        "type": "string"
                      },
                      "image": {
                        "type": "string"
                      },
                      "items": {
                        "type": "array",
                        "items": {}
                      }
                    },
                    "required": [
                      "type"
                    ],
                    "additionalProperties": false
                  }
                },
                "seo": {
                  "type": "object",
                  "properties": {
                    "title": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    }
                  },
                  "additionalProperties": false
                }
              },
              "additionalProperties": false
            }
          }
        ],
        "operationId": "UpdatePageController"
      }
    },
    "/api/v1/cms/pages/{id}/publish": {
      "post": {
        "tags": [
          "CMS"
        ],
        "summary": "Publish page",
        "description": "Publish or unpublish",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "data": {
                  "type": "object"
                },
                "message": {
                  "type": "string",
                  "example": "Success"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "error"
                },
                "data": {
                  "type": "object"
                },
                "message": {
                  "type": "string",
                  "example": "Bad Request"
                },
                "errors": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "error"
                },
                "data": {
                  "type": "object"
                },
                "message": {
                  "type": "string",
                  "example": "Not Found"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "error"
                },
                "data": {
                  "type": "object"
                },
                "message": {
                  "type": "string",
                  "example": "Internal Server Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "publish": {
                  "type": "boolean"
                }
              },
              "required": [
                "publish"
              ],
              "additionalProperties": false
            }
          }
        ],
        "operationId": "PublishPageController"
      }
    },
...
  }
}